feat: auth middleware
